# contracts/mobile/mobile-v1.openapi.json
{
  "components": {
    "schemas": {
      "MobileChargingSession": {
        "properties": {
          "connector_id": {
            "format": "uuid",
            "title": "Connector Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "energy": {
            "$ref": "#/components/schemas/MobileMetric"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          }
        },
        "required": [
          "id",
          "connector_id",
          "vehicle_id",
          "started_at",
          "ended_at",
          "state",
          "energy",
          "reason_codes"
        ],
        "title": "MobileChargingSession",
        "type": "object"
      },
      "MobileChargingSessions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MobileChargingSession"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "station_id": {
            "format": "uuid",
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "items",
          "next_cursor"
        ],
        "title": "MobileChargingSessions",
        "type": "object"
      },
      "MobileChart": {
        "properties": {
          "aggregation": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Aggregation",
            "type": "object"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "measured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MobileChartPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "received_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "resolution": {
            "enum": [
              "15m",
              "1h",
              "1d"
            ],
            "title": "Resolution",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "source": {
            "default": "telemetry_aggregate",
            "title": "Source",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "station_id": {
            "format": "uuid",
            "title": "Station Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Units",
            "type": "object"
          }
        },
        "required": [
          "station_id",
          "timezone",
          "start",
          "end",
          "resolution",
          "units",
          "aggregation",
          "points"
        ],
        "title": "MobileChart",
        "type": "object"
      },
...
      "MobileCost": {
        "properties": {
          "currency": {
            "const": "RON",
            "default": "RON",
            "title": "Currency",
            "type": "string"
          },
          "excludes": {
            "items": {
              "type": "string"
            },
            "title": "Excludes",
            "type": "array"
          },
          "method": {
            "default": "historical_hourly_import_minus_export/v1",
            "title": "Method",
            "type": "string"
          },
          "net_cost": {
            "$ref": "#/components/schemas/MobileMetric"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "system_benefit": {
            "$ref": "#/components/schemas/MobileMetric"
          }
        },
        "required": [
          "net_cost",
          "system_benefit",
          "reason"
        ],
        "title": "MobileCost",
        "type": "object"
      },
...
      "MobileError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/MobileIssue"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "error"
        ],
        "title": "MobileError",
        "type": "object"
      },
...
      "MobileHAContext": {
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/MobileHAEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "insights_consent": {
            "title": "Insights Consent",
            "type": "boolean"
          },
          "integration": {
            "$ref": "#/components/schemas/MobileIntegration"
          },
          "physical_control": {
            "const": false,
            "default": false,
            "title": "Physical Control",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "station_id": {
            "format": "uuid",
            "title": "Station Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "timezone",
          "integration",
          "entities",
          "insights_consent"
        ],
        "title": "MobileHAContext",
        "type": "object"
      },
...
      "MobileIssue": {
        "properties": {
          "code": {
            "enum": [
              "unauthenticated",
              "forbidden",
              "no_station",
              "station_required",
              "integration_disconnected",
              "provider_disconnected",
              "reauth_required",
              "stale",
              "rate_limited",
              "invalid_request",
              "invalid_cursor",
              "service_unavailable"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "MobileIssue",
        "type": "object"
      },
...
      "MobileNotifications": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MobileNotification"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "station_id": {
            "format": "uuid",
            "title": "Station Id",
            "type": "string"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "station_id",
          "items",
          "unread_count",
          "next_cursor"
        ],
        "title": "MobileNotifications",
        "type": "object"
      },
      "MobileOverview": {
        "properties": {
          "costs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MobileCost"
            },
            "propertyNames": {
              "enum": [
                "today",
                "month"
              ]
            },
            "title": "Costs",
            "type": "object"
          },
          "energy": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MobilePeriod"
            },
            "propertyNames": {
              "enum": [
                "today",
                "month"
              ]
            },
            "title": "Energy",
            "type": "object"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "forecast": {
            "$ref": "#/components/schemas/MobileForecast"
          },
          "integrations": {
            "items": {
              "$ref": "#/components/schemas/MobileIntegration"
            },
            "title": "Integrations",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/MobileIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "live": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MobileMetric"
            },
            "title": "Live",
            "type": "object"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "sources": {
            "$ref": "#/components/schemas/MobileSourceSummary"
          },
          "station": {
            "$ref": "#/components/schemas/MobileStation"
          },
          "unread_notifications": {
            "title": "Unread Notifications",
            "type": "integer"
          }
        },
        "required": [
          "evaluated_at",
          "station",
          "live",
          "energy",
          "forecast",
          "costs",
          "unread_notifications",
          "sources",
          "integrations",
          "issues"
        ],
        "title": "MobileOverview",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "ems_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Station-scoped human-session reads. Decimal values are strings; missing values remain null. See contracts/mobile/README.md.",
    "title": "EMS mobile read API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/mobile/charging-sessions": {
      "get": {
        "operationId": "charging_sessions_api_v1_mobile_charging_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileChargingSessions"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Charging Sessions",
        "tags": [
          "mobile-v1"
        ]
      }
    },
    "/api/v1/mobile/charts/energy": {
      "get": {
        "description": "Energy sums and mean SOC from retained rollups; maximum 768 buckets, no raw samples.",
        "operationId": "chart_api_v1_mobile_charts_energy_get",
        "parameters": [
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "24h",
              "enum": [
                "24h",
                "7d",
                "30d",
                "1y"
              ],
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "15m",
                    "1h",
                    "1d"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resolution"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileChart"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Private chart unchanged after reauthorization"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Chart",
        "tags": [
          "mobile-v1"
        ]
      }
    },
    "/api/v1/mobile/home-assistant/context": {
      "get": {
        "operationId": "context_api_v1_mobile_home_assistant_context_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "home_assistant_bridge",
              "enum": [
                "home_assistant_bridge",
                "home_assistant_mqtt"
              ],
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileHAContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Context",
        "tags": [
          "mobile-v1"
        ]
      }
    },
    "/api/v1/mobile/notifications": {
      "get": {
        "operationId": "notifications_api_v1_mobile_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unread",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileNotifications"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Notifications",
        "tags": [
          "mobile-v1"
        ]
      }
    },
    "/api/v1/mobile/overview": {
      "get": {
        "description": "Single first-screen read. Optional station_id auto-selects only a sole accessible station.",
        "operationId": "overview_api_v1_mobile_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Private representation unchanged after reauthorization"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "mobile-v1"
        ]
      }
    }
  }
}